refactor(cargo): split native binstall module
